+++ package.json
@@ -2,7 +2,7 @@
   "name": "project-tracking-database",
   "version": "1.0.0",
   "description": "CS 370",
-  "main": "PDT.js",
+  "main": "PTD.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
